# MM-41085: Use a counting semaphore in sendNotifications