@agnivade 2022-02-04
This is a bit hard to write a unit test for, since we have to verify the number of goroutines currently in-flight and that's a hazy number. But I have tested manually by adding sleep and verified that it works. Please let me know if there's a way to write a test, and I can do that :)