@koox00 2022-02-04
Don't know how we can unit test it, but we could test the results in an api test