@agnivade 2022-02-04
> Don't know how we can unit test it, but we could test the results in an api test

I mean that would verify the correctness (which is not part of this PR), but wouldn't test the semaphore concurrency. There lies the challenge.