@agnivade 2022-02-08
Ok, so in my highly unscientific test, I did not notice any performance drop. Actually, all of the goroutines return their errors in a channel, and then later we range from a slice of `mentionAutofollowChans` to check if any of them had an error. Therefore, essentially, we start all of the goroutines and wait for them to finish. This means we are overloading the DB at once, and waiting for the responses to come. With the semaphore approach, we are throttling the requests at a set concurrency of 8. This does not overload the DB.

I wrote this test to do a rough calculation of how much time it takes:
```go
func TestSendNotificationsSemaphore(t *testing.T) {
	th := Setup(t).InitBasic()
	defer th.TearDown()

	th.App.UpdateConfig(func(cfg *model.Config) {
		*cfg.ServiceSettings.ThreadAutoFollow = true
		*cfg.ServiceSettings.CollapsedThreads = model.CollapsedThreadsDefaultOn
		*cfg.TeamSettings.MaxUsersPerTeam = 500
		*cfg.EmailSettings.SendEmailNotifications = false
	})

	rootPost, appErr1 := th.App.CreatePostMissingChannel(th.Context, &model.Post{
		UserId:    th.BasicUser.Id,
		ChannelId: th.BasicChannel.Id,
		Message:   "rootPost",
	}, true)
	require.Nil(t, appErr1)

	users := []*model.User{}
	for i := 0; i < 100; i++ {
		user := th.CreateUser()
		th.LinkUserToTeam(user, th.BasicTeam)
		th.App.AddUserToChannel(user, th.BasicChannel, false)

		post := &model.Post{
			UserId:    user.Id,
			ChannelId: th.BasicChannel.Id,
			Message:   "hello from " + user.Id,
			RootId:    rootPost.Id,
		}
		_, appErr := th.App.CreatePost(th.Context, post, th.BasicChannel, false, false)
		require.Nil(t, appErr)
		users = append(users, user)
	}

	now := time.Now()
	post := &model.Post{
		UserId:    th.BasicUser.Id,
		ChannelId: th.BasicChannel.Id,
		Message:   "@channel how " + th.BasicUser.Id,
		RootId:    rootPost.Id,
	}
	_, appErr := th.App.CreatePost(th.Context, post, th.BasicChannel, false, false)
	t.Log("time taken: ", time.Since(now))
	require.Nil(t, appErr)
}

```

And both in master and this branch, it takes roughly 800ms.

Here are my results: 
Optimized: 
- 761ms
- 829ms

Master
- 830ms

So it seems like a good improvement without any latency hit.